Добавлена Модуляция
